13-model_state_delete_a.py

+++ 0x0F-python-object_relational_mapping/13-model_state_delete_a.py
@@ -19,6 +19,6 @@ if __name__ == "__main__":
     session = Session()
 
     for state in session.query(State):
-        is "a" in state.name:
+        if "a" in state.name:
             session.delete(state)
     session.commit()
